Переместил из Header.js const navigate, checkAuth, useEffect, формы входа и регистрации в ModalWindow.js удалил старые свои комментарии
